feat(sysv init): Add support for EL6 and SLES11 (#1507)
* Rewritten service file from scratch to account for both RHEL6 and SLES11

* Rearrange order of status so it works correctly on SuSE

* Remove usage of SuSE's recommended startproc, and launch manually

* Make final SuSE compatible changes

* Add LSB headers to service file

* Supporting SysV init via the install script

* Make sed regexp-extended argument more cross platform

* Move environment variables from service file to sysconfig file, where they belong

* Add packaging support for the sysv init file

* Fix incorrect destination for the sysconfig file in the archives

* Remove unneeded exports

* Fixing extraneous information added from copy/paste per PR feedback

* Addressing shellcheck issues

* Additionally corrections for shellcheck recommendations

* Add SysV support to the updater

* Add missing function comments for the new functions

* Update existing tests to pass

* Fix updater error message

* SysV startup script, truncate pidfile on start and add timeout to stop

* Add missing quotes to make shellcheck happy

* Address PR feedback for handling the actual service definitions. Move path components to proper file.

* Remove unneeded error return type

* Update updater linux service tests to be for systemd directly

* Address Joe's experience with starting already running service nuking the pid

* Final (hopefully) pass at integration tests

* Remove old service testing file

* Forgotten _sysv append on go build line for test

* Fix order, and parameter, for EL6 timeout for stop

* Switch to const defines for the service file paths. Remove service name variable usage

* Resolve linter issues

* Update Linux service tests to account for struct changes

* Swallow service commands output in the install script

* Add missing space in echo commands inside the service file

* Implement cleaner LinuxServiceFilePath() code suggestion


* Correct order of chkconfig command exec

* Fix formatting

* fix unit tests

* Add missing newlines to service start, stop, and status commands

* Fix shellcheck issues and systemctl failure in the unix install script

* Implement the suggestion of abusing case statement to become POSIX compatible

* Add a 1 second sleep to prevent the odd failures of restart

---------

+++ scripts/install/install_unix.sh
@@ -19,11 +19,18 @@ set -e
 PACKAGE_NAME="observiq-otel-collector"
 DOWNLOAD_BASE="https://github.com/observIQ/bindplane-agent/releases/download"
 
+# Determine if we need service or systemctl for prereqs
+if command -v systemctl > /dev/null 2>&1; then
+  SVC_PRE=systemctl
+elif command -v service > /dev/null 2>&1; then
+  SVC_PRE=service
+fi
+
 # Script Constants
 COLLECTOR_USER="observiq-otel-collector"
 TMP_DIR=${TMPDIR:-"/tmp"} # Allow this to be overriden by cannonical TMPDIR env var
 MANAGEMENT_YML_PATH="/opt/observiq-otel-collector/manager.yaml"
-PREREQS="curl printf systemctl sed uname cut"
+PREREQS="curl printf $SVC_PRE sed uname cut"
 SCRIPT_NAME="$0"
 INDENT_WIDTH='  '
 indent=""
@@ -66,7 +73,7 @@ fi
 # Helper Functions
 printf() {
   if command -v sed >/dev/null; then
-    command printf -- "$@" | sed -E "$sed_ignore s/^/$indent/g"  # Ignore sole reset characters if defined
+    command printf -- "$@" | sed -r "$sed_ignore s/^/$indent/g"  # Ignore sole reset characters if defined
   else
     # Ignore $* suggestion as this breaks the output
     # shellcheck disable=SC2145
@@ -594,7 +601,7 @@ latest_version()
 {
   curl -sSL -H"Accept: application/vnd.github.v3+json" https://api.github.com/repos/observIQ/observiq-otel-collector/releases/latest | \
     grep "\"tag_name\"" | \
-    sed -E 's/ *"tag_name": "v([0-9]+\.[0-9]+\.[0-9+])",/\1/'
+    sed -r 's/ *"tag_name": "v([0-9]+\.[0-9]+\.[0-9+])",/\1/'
 }
 
 # This will install the package by downloading the archived agent,
@@ -640,17 +647,35 @@ install_package()
     succeeded
   fi
 
-  if [ "$(systemctl is-enabled observiq-otel-collector)" = "enabled" ]; then
-    # The unit is already enabled; It may be running, too, if this was an upgrade.
-    # We'll want to restart, which will start it if it wasn't running already,
-    # and restart in the case that this was an upgrade on a running agent.
-    info "Restarting service..."
-    systemctl restart observiq-otel-collector
-    succeeded
+  if [ "$SVC_PRE" = "systemctl" ]; then
+    if [ "$(systemctl is-enabled observiq-otel-collector)" = "enabled" ]; then
+      # The unit is already enabled; It may be running, too, if this was an upgrade.
+      # We'll want to restart, which will start it if it wasn't running already,
+      # and restart in the case that this was an upgrade on a running agent.
+      info "Restarting service..."
+      systemctl restart observiq-otel-collector > /dev/null 2>&1 || error_exit "$LINENO" "Failed to restart service"
+      succeeded
+    else
+      info "Enabling service..."
+      systemctl enable --now observiq-otel-collector > /dev/null 2>&1 || error_exit "$LINENO" "Failed to enable service"
+      succeeded
+    fi
   else
-    info "Enabling service..."
-    systemctl enable --now observiq-otel-collector > /dev/null 2>&1 || error_exit "$LINENO" "Failed to enable service"
-    succeeded
+    case "$(service observiq-otel-collector status)" in
+      *running*)
+        # The service is running.
+        # We'll want to restart.
+        info "Restarting service..."
+        service observiq-otel-collector restart > /dev/null 2>&1 || error_exit "$LINENO" "Failed to restart service"
+        succeeded
+        ;;
+      *)
+        info "Enabling and starting service..."
+        chkconfig observiq-otel-collector on > /dev/null 2>&1 || error_exit "$LINENO" "Failed to enable service"
+        service observiq-otel-collector start > /dev/null 2>&1 || error_exit "$LINENO" "Failed to start service"
+        succeeded
+        ;;
+    esac
   fi
 
   success "BindPlane Agent installation complete!"
@@ -701,8 +726,13 @@ display_results()
     increase_indent
     info "Agent Home:         $(fg_cyan "/opt/observiq-otel-collector")$(reset)"
     info "Agent Config:       $(fg_cyan "/opt/observiq-otel-collector/config.yaml")$(reset)"
-    info "Start Command:      $(fg_cyan "sudo systemctl start observiq-otel-collector")$(reset)"
-    info "Stop Command:       $(fg_cyan "sudo systemctl stop observiq-otel-collector")$(reset)"
+    if [ "$SVC_PRE" = "systemctl" ]; then
+      info "Start Command:      $(fg_cyan "sudo systemctl start observiq-otel-collector")$(reset)"
+      info "Stop Command:       $(fg_cyan "sudo systemctl stop observiq-otel-collector")$(reset)"
+    else
+      info "Start Command:      $(fg_cyan "sudo service observiq-otel-collector start")$(reset)"
+      info "Stop Command:       $(fg_cyan "sudo service observiq-otel-collector stop")$(reset)"
+    fi
     info "Logs Command:       $(fg_cyan "sudo tail -F /opt/observiq-otel-collector/log/collector.log")$(reset)"
     decrease_indent
 
@@ -750,13 +780,24 @@ uninstall()
   root_check
   succeeded
 
-  info "Stopping service..."
-  systemctl stop observiq-otel-collector > /dev/null || error_exit "$LINENO" "Failed to stop service"
-  succeeded
+  if [ "$SVC_PRE" = "systemctl" ]; then
+    info "Stopping service..."
+    systemctl stop observiq-otel-collector > /dev/null || error_exit "$LINENO" "Failed to stop service"
+    succeeded
 
-  info "Disabling service..."
-  systemctl disable observiq-otel-collector > /dev/null 2>&1 || error_exit "$LINENO" "Failed to disable service"
-  succeeded
+    info "Disabling service..."
+    systemctl disable observiq-otel-collector > /dev/null 2>&1 || error_exit "$LINENO" "Failed to disable service"
+    succeeded
+  else
+    info "Stopping service..."
+    service observiq-otel-collector stop
+    succeeded
+
+    info "Disabling service..."
+    chkconfig observiq-otel-collector on
+    # rm -f /etc/init.d/observiq-otel-collector
+    succeeded
+  fi
 
   info "Removing any existing manager.yaml..."
   rm -f "$MANAGEMENT_YML_PATH"
